+++ src/main/java/net/spell_engine/client/util/Color.java
@@ -53,4 +53,10 @@ public record Color(float red, float green, float blue, float alpha) {
     public static final Color FROST = Color.from(0x66ccff);
     public static final Color ELECTRIC = Color.from(0xffff66);
     public static final Color RAGE = Color.from(0xbf4040);
+
+    public static final Color POISON_LIGHT = Color.from(0x99FF66);
+    public static final Color POISON_MID = Color.from(0x33DD33);
+    public static final Color POISON_DARK = Color.from(0x009900);
+
+    public static final Color BLOOD = Color.from(0xb30000);
 }